+++ src/index.js
@@ -13,9 +13,9 @@ export { polygon } from "./shape/polygon.js";
 export { point } from "./shape/point.js";
 export { linear as scaleLinear } from "./scale/linear.js";
 export { sqrt as scaleSqrt } from "./scale/sqrt.js";
-export { canvas } from "./renderer/canvas.js";
-export { webgl } from "./renderer/webgl.js";
-export { terminal } from "./renderer/terminal.js";
+export { canvas } from "./canvas/index.js";
+export { webgl } from "./webgl/index.js";
+export { terminal } from "./terminal/index.js";
 export { radian } from "./math/radian.js";
 export { clamp } from "./math/clamp.js";
 export { random } from "./math/random.js";
